TST: collect Interval tests
